+++ packages/frameworks/solid/src/rating-group/rating-group-item-context.ts
@@ -1,7 +1,8 @@
 import { type ItemState } from '@zag-js/rating-group'
+import type { Accessor } from 'solid-js'
 import { createContext } from '../create-context'
 
-export type RatingGroupItemContext = ItemState
+export type RatingGroupItemContext = Accessor<ItemState>
 
 export const [RatingGroupItemProvider, useRatingGroupItemContext] =
   createContext<RatingGroupItemContext>({